Restart if any files have changed

+++ GUI/sender_update_script.py
@@ -3,10 +3,21 @@
 import importlib
 import os
 import sys
+from typing import List
 import update_function
 
 
-def run():
+def run() -> List[str]:
+    """Runs update loop and reloads update_function module
+    to add new files avoiding unecessary reboot
+
+    Raises:
+        ConnectionError: IFF internet access was not aquired
+        within certain timeframe (Default 20s)
+
+    Returns:
+        List[str]: List of all files which had been changed or added
+    """
     importlib.reload(update_function)
     from update_function import Updater, SENDER_FILES
 
@@ -22,9 +33,10 @@ def run():
 
 changed_files = run()
 if "update_function.py" in changed_files:
-    run()  # Runs again with new module content
-    if sys.platform == "linux":
-        os.system("sudo reboot")
+    changed_files += run()  # Runs again with new module content
+
+if changed_files and sys.platform == "linux":
+    os.system("sudo reboot")
 
 
 print("Exiting")
